requirements.txt
django
django-mysql
imageio
numpy
pillow
